+++ src/biz/bokhorst/xprivacy/XLocationClient.java
@@ -1,19 +1,17 @@
 package biz.bokhorst.xprivacy;
 
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.Method;
+import java.lang.reflect.Proxy;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.WeakHashMap;
 
 import android.location.Location;
 import android.os.Binder;
 import android.util.Log;
 
-import com.google.android.gms.location.LocationListener;
-
 public class XLocationClient extends XHook {
 	private Methods mMethod;
-	private static final Map<LocationListener, XLocationListener> mListener = new WeakHashMap<LocationListener, XLocationListener>();
 
 	private XLocationClient(Methods method, String restrictionName) {
 		super(restrictionName, method.name(), String.format("GMS.%s", method.name()));
@@ -35,8 +33,6 @@ public class XLocationClient extends XHook {
 	// Location getLastLocation()
 	// void removeGeofences(List<String> geofenceRequestIds, LocationClient.OnRemoveGeofencesResultListener listener)
 	// void removeGeofences(PendingIntent pendingIntent, LocationClient.OnRemoveGeofencesResultListener listener)
-	// void removeLocationUpdates(LocationListener listener)
-	// void removeLocationUpdates(PendingIntent callbackIntent)
 	// void requestLocationUpdates(LocationRequest request, PendingIntent callbackIntent)
 	// void requestLocationUpdates(LocationRequest request, LocationListener listener)
 	// void requestLocationUpdates(LocationRequest request, LocationListener listener, Looper looper)
@@ -45,7 +41,7 @@ public class XLocationClient extends XHook {
 	// @formatter:on
 
 	private enum Methods {
-		addGeofences, getLastLocation, removeGeofences, removeLocationUpdates, requestLocationUpdates
+		addGeofences, getLastLocation, removeGeofences, requestLocationUpdates
 	};
 
 	public static List<XHook> getInstances() {
@@ -53,7 +49,6 @@ public class XLocationClient extends XHook {
 		listHook.add(new XLocationClient(Methods.addGeofences, PrivacyManager.cLocation).optional());
 		listHook.add(new XLocationClient(Methods.getLastLocation, PrivacyManager.cLocation).optional());
 		listHook.add(new XLocationClient(Methods.removeGeofences, null, 1).optional());
-		listHook.add(new XLocationClient(Methods.removeLocationUpdates, null, 1).optional());
 		listHook.add(new XLocationClient(Methods.requestLocationUpdates, PrivacyManager.cLocation).optional());
 		return listHook;
 	}
@@ -71,12 +66,14 @@ public class XLocationClient extends XHook {
 		} else if (mMethod == Methods.getLastLocation) {
 			// Do nothing
 
-		} else if (mMethod == Methods.removeLocationUpdates) {
-			removeLocationListener(param);
-
 		} else if (mMethod == Methods.requestLocationUpdates) {
-			if (isRestricted(param))
-				replaceLocationListener(param);
+			if (isRestricted(param)) {
+				ClassLoader cl = param.thisObject.getClass().getClassLoader();
+				Class<?> ll = Class.forName("com.google.android.gms.location.LocationListener", false, cl);
+				InvocationHandler ih = new OnLocationChangedHandler(Binder.getCallingUid(), param.args[1]);
+				Object proxy = Proxy.newProxyInstance(cl, new Class<?>[] { ll }, ih);
+				param.args[1] = proxy;
+			}
 
 		} else
 			Util.log(this, Log.WARN, "Unknown method=" + param.method.getName());
@@ -92,9 +89,6 @@ public class XLocationClient extends XHook {
 			if (location != null && isRestricted(param))
 				param.setResult(PrivacyManager.getDefacedLocation(Binder.getCallingUid(), location));
 
-		} else if (mMethod == Methods.removeLocationUpdates) {
-			// Do nothing
-
 		} else if (mMethod == Methods.requestLocationUpdates) {
 			// Do nothing
 
@@ -102,58 +96,19 @@ public class XLocationClient extends XHook {
 			Util.log(this, Log.WARN, "Unknown method=" + param.method.getName());
 	}
 
-	private void replaceLocationListener(XParam param) throws Throwable {
-		if (param.args.length >= 2 && param.args[1] != null
-				&& LocationListener.class.isAssignableFrom(param.args[1].getClass())) {
-			if (!(param.args[1] instanceof XLocationListener)) {
-				LocationListener listener = (LocationListener) param.args[1];
-				if (listener != null) {
-					XLocationListener xListener;
-					synchronized (mListener) {
-						xListener = mListener.get(listener);
-						if (xListener == null) {
-							xListener = new XLocationListener(Binder.getCallingUid(), listener);
-							mListener.put(listener, xListener);
-							Util.log(this, Log.WARN,
-									"Added count=" + mListener.size() + " uid=" + Binder.getCallingUid());
-						}
-					}
-					param.args[1] = xListener;
-				}
-			}
-		} else
-			param.setResult(null);
-	}
-
-	private void removeLocationListener(XParam param) {
-		if (param.args.length >= 1 && param.args[0] != null
-				&& LocationListener.class.isAssignableFrom(param.args[0].getClass())) {
-			LocationListener listener = (LocationListener) param.args[0];
-			synchronized (mListener) {
-				XLocationListener xlistener = mListener.get(listener);
-				if (xlistener != null) {
-					param.args[0] = xlistener;
-					Util.log(this, Log.WARN, "Removed count=" + mListener.size() + " uid=" + Binder.getCallingUid());
-				}
-			}
-		} else
-			param.setResult(null);
-	}
-
-	private class XLocationListener implements LocationListener {
+	private class OnLocationChangedHandler implements InvocationHandler {
 		private int mUid;
-		private LocationListener mLocationListener;
+		private Object mTarget;
 
-		public XLocationListener(int uid, LocationListener locationListener) {
+		public OnLocationChangedHandler(int uid, Object target) {
 			mUid = uid;
-			mLocationListener = locationListener;
+			mTarget = target;
 		}
 
-		@Override
-		public void onLocationChanged(Location location) {
-			if (location != null)
-				location = PrivacyManager.getDefacedLocation(mUid, location);
-			mLocationListener.onLocationChanged(location);
+		public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
+			if ("onLocationChanged".equals(method.getName()))
+				args[0] = PrivacyManager.getDefacedLocation(mUid, (Location) args[0]);
+			return method.invoke(mTarget, args);
 		}
 	}
 }