+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Changelog
 * Added restriction to revoke permissions, including white/blacklist
 * Moved *android.intent.action.NEW_OUTGOING_CALL* and *CallLogProvider* from *Phone* to *Calling* category
 * Moved on demand restricting details to expert mode
+* Fixed fake location for Play services
 * Fixed fake location from different application in some cases
 * Updated Dutch translation
 * Updated Italian translation